+++ src/main/java/roomescape/controller/UserController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.User;
-import roomescape.dto.user.CreateUserRequest;
+import roomescape.dto.request.CreateUserRequest;
 import roomescape.service.UserService;
 
 @RestController
