+++ src/main/java/roomescape/dto/command/CancelReservationCommand.java
@@ -1,4 +1,4 @@
-package roomescape.dto.reservation;
+package roomescape.dto.command;
 
 public record CancelReservationCommand(
         Long reservationId,
