+++ src/main/java/roomescape/dto/response/ReservationWithStatusResponses.java
@@ -1,4 +1,4 @@
-package roomescape.dto.reservation;
+package roomescape.dto.response;
 
 import java.util.List;
 import java.util.Map;
