+++ src/main/java/roomescape/dto/response/ThemeReservationTimeResponses.java
@@ -1,4 +1,4 @@
-package roomescape.dto.theme;
+package roomescape.dto.response;
 
 import java.util.List;
