+++ src/main/java/roomescape/dto/response/ThemeResponses.java
@@ -1,4 +1,4 @@
-package roomescape.dto.theme;
+package roomescape.dto.response;
 
 import java.util.List;
 import roomescape.domain.Theme;
