+++ src/main/java/roomescape/service/ReservationService.java
@@ -10,10 +10,10 @@ import roomescape.domain.ReservationTime;
 import roomescape.domain.Store;
 import roomescape.domain.Theme;
 import roomescape.domain.User;
-import roomescape.dto.reservation.CancelReservationCommand;
-import roomescape.dto.reservation.CreateReservationCommand;
-import roomescape.dto.reservation.ReservationWithStatusResponses;
-import roomescape.dto.reservation.UpdateReservationCommand;
+import roomescape.dto.command.CancelReservationCommand;
+import roomescape.dto.command.CreateReservationCommand;
+import roomescape.dto.response.ReservationWithStatusResponses;
+import roomescape.dto.command.UpdateReservationCommand;
 import roomescape.exception.DuplicateReservationException;
 import roomescape.exception.DuplicateWaitingReservationException;
 import roomescape.exception.PastDateTimeReservationException;
