+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -4,8 +4,8 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.ReservationTime;
-import roomescape.dto.reservationtime.CreateReservationTimeRequest;
-import roomescape.dto.reservationtime.ReservationTimeResponses;
+import roomescape.dto.request.CreateReservationTimeRequest;
+import roomescape.dto.response.ReservationTimeResponses;
 import roomescape.exception.ReservationTimeInUseException;
 import roomescape.exception.ResourceNotFoundException;
 import roomescape.repository.ReservationRepository;
