+++ src/test/java/roomescape/controller/ReservationControllerTest.java
@@ -26,9 +26,9 @@ import org.springframework.context.annotation.Import;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
-import roomescape.dto.reservation.CreateReservationCommand;
-import roomescape.dto.reservation.ReservationWithStatusResponses;
-import roomescape.dto.reservation.UpdateReservationCommand;
+import roomescape.dto.command.CreateReservationCommand;
+import roomescape.dto.response.ReservationWithStatusResponses;
+import roomescape.dto.command.UpdateReservationCommand;
 import roomescape.exception.DuplicateReservationException;
 import roomescape.exception.DuplicateWaitingReservationException;
 import roomescape.exception.PastDateTimeReservationException;
