+++ src/test/java/roomescape/controller/ThemeControllerTest.java
@@ -16,8 +16,8 @@ import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 import roomescape.domain.PopularTheme;
 import roomescape.domain.Theme;
-import roomescape.dto.theme.ThemeReservationTimeResponse;
-import roomescape.dto.theme.ThemeResponses;
+import roomescape.dto.response.ThemeReservationTimeResponse;
+import roomescape.dto.response.ThemeResponses;
 import roomescape.infrastructure.AdminAuthorizationInterceptor;
 import roomescape.infrastructure.LoginCheckInterceptor;
 import roomescape.infrastructure.LoginUserArgumentResolver;
