+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -15,8 +15,8 @@ import roomescape.domain.ReservationStatus;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
 import roomescape.domain.User;
-import roomescape.dto.reservation.ReservationResponses;
-import roomescape.dto.reservation.ReservationWithStatusResponses;
+import roomescape.dto.response.ReservationResponses;
+import roomescape.dto.response.ReservationWithStatusResponses;
 import roomescape.exception.DuplicateReservationException;
 import roomescape.exception.DuplicateWaitingReservationException;
 import roomescape.exception.PastDateTimeReservationException;
