+++ src/test/java/roomescape/service/ReservationTimeServiceTest.java
@@ -12,8 +12,8 @@ import roomescape.domain.ReservationStatus;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
 import roomescape.domain.User;
-import roomescape.dto.reservationtime.CreateReservationTimeRequest;
-import roomescape.dto.reservationtime.ReservationTimeResponses;
+import roomescape.dto.request.CreateReservationTimeRequest;
+import roomescape.dto.response.ReservationTimeResponses;
 import roomescape.exception.ReservationTimeInUseException;
 import roomescape.fixture.Fixtures;
 import roomescape.repository.fake.FakeReservationRepository;
